requirements.txt
openai

python-dotenv

google-generativeai

langchain_openai

langchain_google_genai

langchain_ollama

langchain_groq

langchain_anthropic

langchain

langchain_core

langchain_aws

boto3